+++ lib/src/implementation/duckdb.dart
@@ -630,6 +630,14 @@ class Connection {
                       isUtc: true);
             }));
 
+          case DUCKDB_TYPE.DUCKDB_TYPE_TIMESTAMP_TZ: // return the number of microseconds
+            var xs = values.cast<Int64>();
+            out[name]!.addAll(ids.map((i) {
+              final isNull =
+                  !bindings.duckdb_validity_row_is_valid(validity, i);
+              return isNull ? null : xs[i];
+            }));
+
           case DUCKDB_TYPE
                 .DUCKDB_TYPE_TIMESTAMP_MS: // UTC DateTime, millisecond precision
             var xs = values.cast<Int64>();