+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - Implement appender
 - Tests break with ffigen 14.0.0.  Has a different enum treatment. 
 
+## 2024-10-04 (0.3.5)
+- Map TIMESTAMPTZ data type to Dart.  Return the number of microseconds since epoch. 
+  Would be nice to have a database that natively supports timezones, eh?
+
 
 ## 2024-09-15 (0.3.4)
 - Introduce `fetchRows` to map the result of a DuckDB query to a Dart class.